Struct almonds::Almond

pub struct Almond {
    // some fields omitted
}

A representation of a deserialized Almond.

Every Almond consists of:

A generation, which is a simple version to allow upgrading the caveat formats. This is an 8-bit unsigned integer.

A type, which defines what this caveat is used for. For example, general access vs. logging in a new client.

A list of caveats that each decrease the level of authorization the Almond allows

The type and caveats are serialized within the almond, so it is suggested that they are not needlessly verbose.

A caveat is key (with optional value) that decreases the scope of the Almond. An almond with no caveats grants full authorization to any and all holders of the Almond. In practice, almost all almonds have at least one caveat that scopes it to a particular user.

The exact format and interpretation of the caveats are application defined.

Methods

impl Almond

fn create(key: &[u8], generation: u8, almond_type: Vec<u8>) -> Almond

Create a new Almond with given generation and type.

fn parse_and_validate(key: &[u8], input: &[u8]) -> Result<Almond, AlmondParseError>

Parse a binary serialized Almond, and validate that the hashes match.

Note: This expects a binary serialization rather than base64

fn parse_base64_and_validate(key: &[u8], input: &[u8]) -> Result<Almond, AlmondParseError>

Parse a Base64 serialized Almond, and validate that the hashes match.

fn add_literal_caveat(&mut self, caveat: Vec<u8>) -> &mut Self

Add a new literal caveat.

The interpretation of the caveat is either <key> or <key> <value> depending on if caveat has a space or not.

fn add_caveat(&mut self, key: &[u8], value: Option<&[u8]>) -> &mut Self

Adds a caveat.

The key must not include a space.

fn almond_type(&self) -> &[u8]

Get the type of the Almond

fn generation(&self) -> u8

Get the generation of the Almond

fn caveats(&self) -> &[Vec<u8>]

Get the current caveats of the Almond

fn hash(&self) -> &[u8; 32]

Get the current hash of the almond.

Safety

Do not compare this directly with other hashes. Always use a specially designed constant time comparison function.

fn serialize_binary(&self) -> Vec<u8>

Serialize into a binary blob

fn serialize_base64(&self) -> String

Serialize into Base64.

This is equivalent to Base64 encoding the binary serialization