Struct almonds::Verifier

pub struct Verifier<'a> {
    // some fields omitted
}

The verifier takes an almond and checks if it satisfies a list of predicates.

An almond will be accepted if:

All caveats match and pass at least one predicate.

No caveats are rejected by any predicate.

In particular, an almond will be rejected if it has any "unrecognized" caveats, i.e. ones that do not match any predicates. On the other hand, not all predicates must have matched a caveat.

Methods

impl<'a> Verifier<'a>

fn new(almond: &'a Almond, generation: u8, almond_type: &[u8]) -> Verifier<'a>

Create a new instance to verify the given caveat.

fn allow(&mut self, key: &[u8]) -> &mut Self

Allow all caveats with the given key, irrespective of their values.

This should generally only be used if the value will be pulled out of the almond for usage.

Note: If you are expecting a key with no value then use satisfies_exact with value set to None

let mut almond = Almond::create(b"secret", 1, b"access".to_vec());
almond.add_caveat(b"user", Some(b"erikj"));

let mut v = Verifier::new(&almond, 1, b"access");
v.allow(b"user");
assert!(v.verify());

fn satisfies<F>(&mut self, key: &[u8], predicate: F) -> &mut Self where F: FnMut(&[u8]) -> bool

Invokes predicate against the value of every caveat with the given key. If predicate returns true then the caveat is accepted, false rejects the caveat.

Rejects the caveat if the key matches but has no value.

let mut almond = Almond::create(b"secret", 1, b"access".to_vec());
almond.add_caveat(b"expires", Some(b"1500000000"));

let mut v = Verifier::new(&almond, 1, b"access");
v.satisfies(
   b"expires",
   |val| {
        str::from_utf8(val).ok().and_then(
            |val| val.parse().ok()
        ).map(
            |val| 1447720058 < val   // Where `1447720058` is 'now'
        ).unwrap_or(false)
    }
);
assert!(v.verify());

fn satisfies_exact(&mut self, key: &[u8], value: Option<&[u8]>)

Compares value with the value of every caveat with the given key. If they match then the caveat is accpeted, otherwise it is rejected.

let mut almond = Almond::create(b"secret", 1, b"access".to_vec());
almond.add_caveat(b"user", Some(b"erikj"));

let mut v = Verifier::new(&almond, 1, b"access");
v.satisfies_exact(b"user", Some(b"erikj"));
assert!(v.verify());

fn verify(&self) -> bool

Returns whether the almond satisfies the given conditions and whether all caveats have been accepted by at least one condition.

Always returns false if the almond does not have match specified almond_type and generation.